backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.0
httpx==0.27.0
python-docx==1.1.0
ebooklib==0.18
beautifulsoup4==4.12.2
mobi==0.3.3
PyPDF2==3.0.1
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
pydantic[email]